refactor: move cmd/ghx to root level cmd
