+++ cmd/ghx/root.go
@@ -9,13 +9,13 @@ import (
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
 
+	"github.com/aweris/gale/cmd/ghx/run"
+	"github.com/aweris/gale/cmd/ghx/version"
 	"github.com/aweris/gale/internal/config"
 	"github.com/aweris/gale/internal/core"
 	"github.com/aweris/gale/internal/fs"
 	"github.com/aweris/gale/internal/journal"
 	"github.com/aweris/gale/internal/log"
-	"github.com/aweris/gale/tools/ghx/cmd/ghx/run"
-	"github.com/aweris/gale/tools/ghx/cmd/ghx/version"
 )
 
 // NewCommand  creates a new root command.
